Add support for defining a block distribution without a bounding box
This extends the block distribution's initializers to support a mode
in which the distribution need not be given a bounding box, but will
infer it from the first domain declared over it.  Note that due to
expression evaluation order, it still must be given a rank.  Here I've
done that by adding an initializer that takes just a rank; and another
that takes a targetLocales array and infers the rank from that.

The approach taken here is to use an empty 'boundingBox' field as an
indication that one of these initializers was used, and to use that as
a signal to grab the first domain declared over it.  Though that could
be considered a bit weak since it would permit someone to redistribute
the distribution to an empty bbox and then have it re-grab its
bounding box from the next domain declared over it, which seems
potentially confusing.  Or is that a benefit?  Another approach would
be to store a bit-field making this happen just the first time when
one of those two initializers is used.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>