Update chpldoc to change the ordering of deprecation/unstable msgs w.r.t docs
We want deprecation and unstable documentation notes to be generated before the
documentation comment.  When it was at the end of the documentation comment, it
could confuse users into thinking that it was associated with the later symbol
instead of the preceding one.  By placing it between the symbol header and its
documentation comment, it's more clear what it is associated with.

While here, ended up mucking about a bit with the spacing after class comments
in general.  I think the situation is improved now, but it will require updating
some test files

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>