Fix modules to use the updated ordering with deprecation/unstable, too
Modules handle most of their printing in a different code path than other
symbols, in part because they print out additional boilerplate information.
However, this meant that spot needed a separate update.

As a result of that change, the extra newlines were not actually necessary for
some reason, so remove their special casing from the unstable handling (need to
do it for deprecation handling as well, but I don't have a test to lock it in
yet)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>