Fix assumption in LinearAlgebra utility routine that domains were DR 2D
This fixes an assumption in the LinearAlgebra helper routine,
hasNonStridedIndices(), that the domain being queried is a 2D
DR routine.  Here, I've broadened it to support any nD domain
and added some param overloads for cases where true or false
can trivially be returned at compile-time.

Currently, this routine is only used by the dot() routine when it's
constrained to 2D arrays.  However, those 2D arrays could be 2D arrays
of things other than default rectangular domains, such as 2D
rank-change slices of a larger array.  By generalizing the routine in
this way, we permit such domains to be passed in.  We could also
constrain the domain's rank to be 2, but there doesn't seem to be any
need to do this at this time, and the more general version may be
useful later.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>