Modularize gpu scan functions and add testing
Adds a gpuScan proc to the GPU module library.
This uses a few other functions under the hood in order to execute a
scan on the GPU using a few heuristics to decide which one to use.
Currently this only works for uints, but there is nothing keeping us
from expanding support to other integral types as well,
and maybe even other number types.

Need to fix documentation and remove duplicate code from radix sort

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>