Add cast to tmpdirname
This adds a type cast to tmpdirname as a means of trying to make
error messages go away when using the --gdb flag that complain
that the type can't be determined, suggesting to apply a cast.

This change is a bit speculative because I haven't been able to
reproduce the error, have just seen that it occurs in some
configurations and that we have tests that try to grep the
message away.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>