Convert gasnet-ex shim strided get/put over to GEX API
Convert strided RMA calls over to GEX API. The strided APIs were made a
little more general in GEX, so adapt to the changes here. Specifically,
`count[0]` is no longer special and was moved out into a separate
`elemsz` variable so `count` is now of `stridelevels` size  instead of
`stridelevels+1`.

Our strided interface was at least partially based on the gasnet-1
interface, so I think it could be worth revisiting the interface to see
if it could be simplified.

Note that during review we discovered a bug where we were using 0 length
VLAs for `stridelevels == 0` (contiguous transfers). 0-size VLAs are
undefined-behavior, so fix that here by ensuring the arrays are always
at least of size 1 (they will be ignored by gasnet for stridelevels==0).
I suspect in practice this didn't matter since clang and gcc have
extensions for 0-length VLA, but it's still UB (and is caught by UBSan)
so we should backport the fix to the gasnet-1 interface.

Part of 14443

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>