Move check out of NDEBUG #ifndef since it is now user-facing...
[I hadn't noticed this before]

Interestingly, failing to do this causes a NDEBUG build to issue a
standard "Redefinition of..." error further down in compilation.  Yet
it prints the same location twice which makes it not useful.  Catching
the error early here seemed more attractive than trying to fix that
code given that Dyno will take this over at some point (at which point
perhaps this check can be removed).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>