Temporary change to perftesting pending bug fix
There is an internal error caused when calling out to any kernel in that
resides in `modules/standard` and therefore sorts and scans need to use
the `--devel` flag as a workaround.
Since this isn't amenable to perf testing, I'm removing it for now and
I'll add it back in once this bug is fixed.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>