Rework locale setup to avoid dummyLocale
This is some code massaging I've been wrestling with to avoid the
problem of creating a dummy locale and associating certain key classes
(like stdin, stdout, etc.) with it rather than locale 0.  The general
idea is to allocate locale 0 instead of dummy locale and then to poke
it into the locales array when we're setting that up rather than
creating a new locale 0 value.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>