Update passCheck.prediff to account for output order change
I understand this contradicts the comment at the top saying not to
update this prediff if the test is failing. However, I believe in this
case the test really is just catching lines the perfkeys would ignore
(due to the desired total times now being printed after everything else),
so I think it is the correct solution.

As an additional reassurance, this prediff change doesn't cause the test
to fail on main.

Signed-off-by: Anna Rift <anna.rift@hpe.com>