Perform some optimizations to write/readBinary()
This makes an optimization to write/readBinary() by using the
fast-path code whenever the endianness requested explicitly matches
the current architecture's endianness rather than falling into the
slow path for both big/little choices as it has been.

While here, it also removes an unnecessarily complex select statement
from the slow-path code.  Since the endianness is a param, we can
simply use a param routine to convert it to the corresponding ioKind
param rather than having a case per value.

Weirdly, this code is failing on
library/standard/IO/readBinary/arrayOfValues.chpl for reasons that I
haven't been able to get my head around—the failing cases have all
seemed to be in the fast-path branch so far.

Also as a TODO:

- [ ] BHarsh mentioned some other places he thought would need to
      change as well, which I haven't really looked at or digested
      yet, jumping straight into the code instead to see if it was
      as easy as I'd hoped.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>