Skip new "warn about -pg with LLVM" test for valgrind runs
Something about the configurations that use `--ldflags -pg` leads to a
segfault in valgrind testing.  Since this test isn't about valgrind or
support for profiling with valgrind, I'm just skipping it in valgrind
mode.  I considered suppressing it, but only two of the five
configurations fail (the two that use `--ldflags`), and I don't think
it's worth the effort to break into two tests or the time spent
testing under valgrind at this point.  If/when we support profiling as
more of a first-class feature, we can revisit this.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>