Generalize assignment of NICs to co-locales
Generalizes NIC assignment to handle NICs that aren't in a socket. The
assignment is based on the locales' cpusets and distances to NICs, rather than
on locales and NICs sharing sockets. Each locale is represented by the lowest
common ancestor of its cpuset, and each NIC is represented by its lowest
ancestor that has a cpuset. The distance between a locale and a NIC is the
length of the path between them in the topology. The assignment process finds
the shortest path and assigns that NIC to that locale. Then it repeats with
the remaining locales and NICs until all locales have a NIC or there are no
remaining NICs. If some locales do not have a NIC the entire process is
repeated until all locales are assigned a NIC. This heuristic avoids sharing
NICs unless there are are more locales than NICs.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>