Second attempt at skipping this test under valgrind
Though the skipif I committed yesterday worked as expected when
running 'start_test -valgrindexe' locally, it did not get skipped in
nightly testing as expected.  Jade and I both guessed it could have
something to do with the end-of-line comment, so here, I'm hoisting
the comment onto a line of its own.  This also causes the test to
be skipped in local testing; hopefully it will in nightly as well.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>