Move new roundToPowerof2 out of Math module to GPU
Since this function has not recieved any design discussion yet, we don't
want to add it to the Math module (which has been stabilized).
For now, since the only use case is in the GPU module, I'm moving it
back there. We can move it to the Math module once it is stabilized

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>