Fix internal compiler error with classes
Were encountering an issue because classes triggered the creation of an
instantiation of the function and this called `concreteIntentForArg`, causing
us to try an insert the checks when we didn't even have the body of the function
yet.  Since that call to concreteIntentForArg didn't change the argument intent,
delay responding to it until when resolveIntents is run on the function later.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>