Update code example in reduceIntents tech note
It was out of date with recent Chapel changes.

Add "do" to the braceless single-statement functions.
Add a "(?)" to a generic typed argument.

I took the liberty of changing the second example to use different
variable names than the first example so that the whole code block
could be compiled at once.

---
Signed-off-by: Paul Cassella <gitgitgit@manetheren.bigw.org>