Modify ReplicatedDist to use parSafe=false default
This was causing warnings from the change in default parSafe mode, so I
silence these by using parSafe=false. This might cause problems if there
is something that relied on the behavior of parSafe=true. Maybe this
should be looked into more. From a cursory look, it didn't seem
problematic.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>