Fix some test failures, adjust strategy for others.
Fixed some tests that had the warnings for parSafe causing a mismatch in
the good file and output. So far I've looked at each test and seen if it
makes sense to just use parSafe=false to avoid future work when the
warning goes away, or just silence it using -snoParSafeWarning.

After this I will run another paratest to see how many remain (because
many failures were caused by ReplicatedDist). If there are a lot of
failures, I don't want to go through each of them so I will just a
script to silence all warnings by modifying the compopts in each case.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>