Programmatically add compopts for -snoParSafeWarning
Since there were so many test failures caused due to the new warning
with associative domains, it needs to be silenced in many cases.
Therefore I indiscriminately added the -snoParSafeWarning for all tests
that were failing with this failure mode. Will run paratest after this
to see how well it works.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>