Support for one locale per NUMA domain
Add support for one locale per NUMA domain. If the number of locales per node
is greater than the number of sockets, but less than or equal to the number
of NUMA domains, then each locale will be confined to a NUMA domain. The
environment variable CHPL_RT_FORCE_COLOCALE can be used in co-locale testing.
Setting CHPL_RT_FORCE_COLOCALE=socket will confine the locale to a socket
even if there is only one locale on a node. Similarly,
CHPL_RT_FORCE_COLOCALE=numa will confine the locale to a NUMA domain.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>