Support for per-device SS11 authentication
Slingshot-11 devices, service IDs, and VNIs are passed to the locale in the
environment variables SLINGSHOT_DEVICES, SLINGSHOT_SVC_IDS, and
SLINGSHOT_VNIS, respectively. If there are multiple NICs then
SLINGSHOT_DEVICES will be a comma-separated list, in which case
SLINGSHOT_SVC_IDS and SLINGSHOT_VNIS might also be comma-separated lists. If
SLINGSHOT_SVC_IDS is a list, then we should authenticate using the element of
the list that corresponds to the index of our device in the SLINGSHOT_DEVICES
list. SLINGSHOT_VNIS should be handled similarly.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>