Sort NICs based on PCI bus address
The NICs are sorted during NIC selection to ensure that all locales see them
in the same order. Previously, they were sorted based on the hwloc os_index,
which isn't guaranteed to be unique, and furthermore, the arguments to the
comparison routine were dereferenced incorrectly. This should fix both of
those problems, the former by sorting based on the PCI bus address.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>