Improve how tuples are converted into complex values
This changes the implementation of the _chpl_complexNNN() routines in
the runtime to use the 'CMPLX()' initializer macros rather than math
to initialize the complex values.  This seems to work around an issue
in which inf/nan values would creep in for reasons I don't (yet)
understand.

The one trick to this change is that I had to protect the code to not
compile it in C++ mode.  I believe this is fine because re2 is the
only runtime code we compile with C++ and it doesn't need our complex
types and routines.  This almost made me compelled to break the
complex types out of chpltypes.h and into their own header, but that
felt like a lot of trouble to go to for C++ when we're mostly focused
on C in the runtime.

If anyone going forward has suggestions for making these conversion
routines work equally well in C as C++, I'm all ears!

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>