Add CHPL_RT_COMM_OFI_EXPECTED_PROVIDER
The CHPL_RT_COMM_OFI_EXPECTED_PROVIDER environment variable specifies the name
of the provider that should be chosen by the ofi communication layer. It is
an error if the selected provider does not match this value. This variable is
intended to be used during testing to ensure that the correct provider is
chosen.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>