Update some of the Python dependencies for chpldoc
This was motivated by one of the Sphinx dependencies getting updated from under
us and causing extensive chpldoc failures to build.  Sphinx needed to be at
least 5.0, and that meant that breathe needed to be updated to at least 4.34.0.

Unfortunately, breathe currently reports more warnings than we want, so we will
need to do a build update to stop generating so many warnings until the compiler
documentation can be fixed to avoid those warnings.

While here, I'm also updating other dependencies, since that needed doing
anyways.

Updates:
- MarkupSafe from 2.1.1 to 2.1.3
- Sphinx from 4.5.0 to 5.3.0
- docutils from 0.16 to 0.18
- sphinx-rtd-theme from 1.0.0 to 2.0.0
- breathe from 4.31.0 to 4.35.0

Note that this will not build yet - it requires an update to the Chapel Sphinx
domain, which is currently in progress.  Using the domain locally seemed to
behave okay, though.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>