Fix C++ portability issue with using CMPLX
Though compiling PR #24184 with a back-end C++ compiler worked on my laptop
and chapcs11, it caused problems with other C++ compilers, like g++ (SUSE
Linux) 7.5.0 on [redacted] and some of our other nightly configurations.

At present, it's somewhat circumstantial that we compile this code with a C++
compiler: It only happens when compiling our qio re2 shim code, and that code
doesn't rely on these tuple-to-complex conversion routines anyway.  As a
result, what I've done here is to beef up my #ifndef to skip over the new
preferred path of code when we're compiling for C++ to be safe, and on the
assumption that it doesn't really matter anyway (won't be called); and even if
that changes in the future, the fallback paths from #24184 ought to work...

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>