Remove the deprecated versions of log1p and their deprecation test
This has been deprecated for several releases.  We would have removed it sooner,
but had some decisions to make about how to handle the rest of the log
functions.  Now that we know our plan for those functions and they have been
deprecated for multiple releases, it is safe to remove these old versions of
log1p

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>