Revert "Update pycparser version to latest release"
This reverts commit 392d8ba62ee8ca24f45a33f4ed9487e6ad4c77de.

Despite this seeming like a reasonable update, this caused c2chapel testing to
break with multiple missing extern declarations.  When I had checked locally, I
had assumed that was because my setup was incorrect, since I still saw the
failures with the previous version of pycparser that had worked.  Weird.

Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>