Downgrade filelock depended on by start_test to a slightly older version
The version I grabbed actually required Python 3.8 (as opposed to some of these
packages which say they do, but don't necessarily need that recent a version of
Python).  Though we can drop our Python 3.7 testing, this is still a more recent
version of filelock than we had been using before, so in the interest of getting
testing back fully (especially since I put off upgrading Sphinx fully), take the
path of least resistance until we have updated the versions of Python we are
testing.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>