Add support for co-locale suffix
Allow the "-nl NxL" argument to have a suffix so it is now of the form
"-nl NxLt". The 't' suffix specifies the type of architectural feature in
which the co-locales should run. An 's' suffix indicates that co-locales
should run on separate sockets, 'n' indicates separate NUMA nodes, 'c'
indicates separate cores, and '$' indicates separate level 3 caches. Thus
specifying "-nl 1x2s" will create two co-locales on a single node, each
running in its own socket. In this instance it is an error if the node has
fewer than two sockets. If the node has more than two sockets the extra
sockets will be unused.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>