Squash warnings about missing '?'s in compiler-generated code
It seems that, at present, we sometimes warn about missing '?'s in
code that the compiler generated, such that there's nothing the user
can do to silence the warnings rather than squashing them.  This PR
checks whether the type expression lacking the '?' was defined in a
compiler-generated routine, and if so it squashes the warning.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>