Modify normalization of domain queries to avoid extra copies for arrays
In #24330 we observed that extra copies were occuring for initializers
written to take an array by in-intent:

```chpl
proc ArrayWrapper.init(in a: [?d] ?t) {
  this.etype = t;
  this.rank = d.rank;
  this.shape = d.shape;
  this.a = a; // extra/unexpected copy here!

  writeln(d); // <- commenting this line out removes the extra copy
}
```

This was surprising since the 'in' intent rules are fairly well-defined
and implemented.

The cause was that the normalize pass was transforming the writeln call
into something like ``writeln(a.domain)``. The late mention of 'a' would
then prevent the split-init logic from eliding the copy.

This commit changes the normalized representation of domain queries to
just use a single 'const ref' temp that we set up at the beginning of
the function.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>