Resolve the valgrind failure for two more cases
The iterator temporary inserted by rebuildIterator was not getting initialized
in the case where it only had a dummy field.  This was causing problems when we
tried to hash the iterator when it was passed as an argument directly.

I could choose to only do this memset when we are going to hash the iterator
argument for the check, but given that the uninitialized memory is being copied
regardless, it might be justified to always do so.  We'll see if there's a
performance impact

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>