Fix the remaining failures (as of last time I checked)
The string type has two boolean fields, so was encountered the same general
problem as the record case.  However, we create strings in an unusual way, so
needed to zero the memory where the value was created - this was in
`chpl_createStringWithBorrowedBufferNV` rather than the compiler.

I intend to guard this zero'ing so it only fires when the warning would, but I
wanted to check in a snapshot of the thing that worked first

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>