Make it possible to check if the constness checking flag is set from the modules
This will support only calling memset on strings when we're being cautious about
indirect modifications, instead of having it happen all the time.  I didn't
check if there was a performance impact that this avoids, but it seemed likely
there would be.

While here, added a comment for why we're doing the memset in the first place

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>