Set CHPL_RT_MAX_HEAP_SIZE
On a HPE Cray EX test machine cxi services sometimes die expectedly. The
probability of this happening appears to be proportional to the size of
the heap. Limit the size of the heap to 16GB via CHPL_RT_MAX_HEAP_SIZE
as a larger heap is not needed by the tests and it reduces the chance of
spurious test failures.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>