Add a test locking in public use of Math working
Issues #24010 and #24494 noted that Math could not be publicly used,
which was due to some of the deprecations that were added as symbols
were moved between Math and AutoMath.  Now that those deprecations
have all been removed, this works, so I'm adding in a test to lock in
the behavior.

Resolves #24010.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>