Don't perform const checks on type arguments, even if they're runtime types
Fixes a bug that we encountered when testing an unstable warning for associative
domains - that test had a type argument which was getting passed a runtime type.
Since runtime types by definition must exist at runtime, the argument had been
transformed and was now relying on the default argument intent, making it
inaccurately eligible for the indirect modification const check.  Something
about the runtime type was causing valgrind to complain, which is how we noticed
these unnecessary checks were getting inserted.

Type variables are only created, they cannot be modified, so it is unnecessary
to worry about if they will be indirectly modified over the course of a function
which takes them as an argument.  So just avoid inserting the checks when we
know we have been given a runtime type info argument.

Verified that the test no longer gets a const check inserted for that argument
but that const checks are still inserted by checking the generated code.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>