Cache results from calls to getConfigParamBool
This is an expensive operation to get the value of a config paramwhich
is linear in the number of params in the ChapelBase module. This means
that we want every place which calls this function to cache the result
for later use. So far, there are only two uses of this function which
cache the result as of this commit.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>