Extend the man page description of `--[no-]const-arg-checks`
Brad pointed out that we didn't have documentation for the `const` and default
argument intent indirect modification warning.  Since the check is
implementation-specific, it doesn't make sense to mention it in the spec.  Brad
suggested the man page as a place people are likely to encounter it in the
steady state (and the entry didn't necessarily give sufficient context to begin
with).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>