Refactor default function helpers to take a QualifiedType
This refactoring allows for the creation of both type and value methods
depending on the way in which the default function was called. With this
functionality available, this commit also allows for calling ``eltType``
on either a c_ptr/Const type or value.

Also adds ``CPtrType::id(Context*)`` to more easily grab the correct ID,
instead of checking ``isConst()`` and using ``getId`` or ``getConstId``.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>