Look for HPE/EX-related showstoppers in cray.html
While reviewing the website for the 2.0 release today, we realized
that HPE Cray EX isn't highlighted as a platform on our download page
alongside XC, which is increasingly being put to bed.  Updating the
website is easy, but most of the material is here in the platform
docs, which was still very XC-centric in terms of ordering of
material.

Here, I've taken a pass to try and bring HPE EX and Apollo more to the
forefront relative to Cray XC and CS, while also touching up some
other things that caught my eye.  I think there's more we could/should
do here to bring HPE systems to the forefront and give them equal
time and attention to the historical Cray systems, but am currently
imagining that to be post-2.0 work.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>