Adjust base Dockerfile to not remove a dependency of derivatives
The gasnet and gasnet-smp Dockerfiles are based off of the base chapel
Docker image. However, with the adjustment to a multi-stage build in
which the test dirs aren't kept around, these builds are failing as we
have a frontend test relying on copying some test data under test/types.
As a bandaid solution, keep that subdirectory of test/ around in the
final stage.