Add a test locking in a bug I found with error handling
I was encountering a bug where a reduction inside a forall loop that also called
a throwing function would trigger an assertion error during error handling.  It
appeared to be happening because the reduction was transformed into a forall,
but the forall wasn't turned into a loop that could throw, so wasn't getting
tracked on the tryStack, but we were still marking it as entered.

This bug was impacting HPO.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>