Simplify check for param formals that aren't truly generic
This branch is encountered by most (all?) param formals because they
will not have a param value yet. We first check to see if we can
instantiate the formal from the actual based on its type expression.
Otherwise, we fall back on the switch statement logic below.

Also adds a relevant test case.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>