Add support for syncWarp
For Nvidia GPUs, starting with Volta architectures and later threads within a
warp can diverge. For applications needing memory and synchronization
barriers, calling syncWarp provides this functionality within warps.
It is like syncThreads, but at a more granular level.

For AMD GPUs, threads within a warp are guaranteed to be executed in
lockstep. Therefore syncWarp is a no-op in that case. This may change in
the future as new AMD architectures emerge.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>