Switch non-locking stdin/stdout to use new idioms and names
After some discussion with the performance team, we decided
we like this way better of creating non-locking stdin/stdout
channels -- essentially cloning the default ones and renaming
them.

(That said, we'd also like either built-in unlocked stdin/stdout
or a simpler .getNonLocking() method that could be applied to
clone any such channel).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>