Restore ': real' type constraints + therefore performance to spectralnorm2
Yesterday, in doing some code cleanups, I couldn't remember why this
version of spectralnorm's 'A' routine took explicit ': real' arguments
and the other took the generic (therefore 'int') arguments.  At the
time, I was thinking it was a correctness issue, and seeing that things
worked fine either way, removed the constraints for cleaner code.

Last night's performance testing reminded me that it's really a matter
of performance, as the 'real' versions permit better vectorization to
take place.

As a result, I'm retoring them here, to restore the performance.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>