Make knucleotide support misses when looking up strings
It seems that the problem size we've been testing (which I believe is
based on the original suggestion for verification) assumed that all
strings would be found, but broke with the new, smaller, verification
size which doesn't include the last few strings.  Here, I changed our
map indexing to use the .get() method, passing in 0 as a sentinel,
which fixes the issue, and I added the new smaller input/output files
to the correctness runs.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>