Adjust some off-by-one issues
In reviewing these, BHarsh asked why the domain had a +1 on it.
Looking into it to remind myself (if I ever knew), the answer is that
the incremental readLine() are each subtracting one from the cursor
position in the buffer to overwrite linefeeds.  However, this +1 at
the end was preserving the final linefeed.

But did we need to preserve the final linefeed?  After all, the rest
of the buffer just contained acgt data.  Turns out, the answer is "no"
as long as we adjust a couple of other expressions to add 1 back in
again.  This makes some of the later expressions incrementally more
complex, but also seems appropriate from a "size = hi-lo+1"
perspective and in the sense of "Why keep a single linefeed at the end
of a data buffer that isn't necessarily intended to be printed out?

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>