c_ptr is no longer considered a builtin type
With c_ptr no longer being a builtin type, we can more directly rely on
the fake "c_ptr" class in CTypes.chpl as an "anchor" for the type's ID.

A few changes to resolution-queries are required to handle some CPtrType
special cases. The main change is to lookupCalledExpr, which allows for
the resolution of methods on c_ptr even when CTypes is no longer in
scope.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>