Change param-based nbody to foreach-based
While thinking about our fastest/smallest nbody entry on yesterday's
CLBG results, I was wondering how a foreach-based version would
compare to the for+param version in terms of performance these days.
It will probably be less compact in terms of code size—particularly
because we're using a tuple of records, which require a `with` clause
to modify—but it also feels like a more intuitive use of the language
features.  And since my study version has been in sync with the
submitted nbody3.chpl, no need to keep timing it as it was.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>