Fix nl argument parsing
CHPL_RT_LOCALES_PER_NODE should be set if there is a single locale per node
that is bound to an architectural feature. Otherwise the runtime will not
partition resources correctly. Updated the test cases to test for this, and
to be more rigorous about when this and other environment variables should
be set and not set.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>