Avoid string copies by implementing replace with match
This commit avoids string copies that were coming from moving the data
to and from a std::string when calling RE2's Replace/GlobalReplace.  It
does so by using our own implementation of Replace by calling RE2's
Match function in the re2-interface.cc code so that we can construct a
result buffer by repeatedly doubling & then provide that when
constructing a Chapel string.

While there, this removes the need for doReplaceAndCountSlow, since now
that we have our own implementation of the replace function, we can
handle a maximum number of matches right there.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>