Fix resolution of module-scope MultiDecl and TupleDecl
This commit fixes a bug in which module-scope MultiDecl or TupleDecl
variables had an UnknownType outside of their declarations. The problem
was that ``typeForModuleLevelSymbol`` was trying to fetch the previously
resolved variable with ``resolveModuleStmt`` using the ID of the
variable. This didn't work because the ID of a MultiDecl or TupleDecl
entry is not at the statement level and so a previous query result did
not exist.

The solution is to use 'idToContainingMultiDeclId' which will give us
the correct module statement ID.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>