Switch aggregation-based bale benchmarks to use Block by default
These benchmarks were originally written to use Cyclic by default in
order to more closely mimic the reference versions that they were
based upon.  In our performance runs and work, we've definitely
focused on Block in recent runs and years, so here I'm switching the
default to indicate that preference without removing the ability to
switch back to Cyclic.  Updated the README by flipping the flags
specified on the 'chpl' command lines.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>